requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

apache-airflow
apache-airflow-providers-microsoft-azure
azure-storage-blob
requests
langdetect
langcodes
apache-airflow-providers-snowflake>=5.0.0
